Adds processing including number of puts
Which reveals a curious fact. In many cases, experiment take a single PUT which means that it is probably a run over from workers running previous experiment. I don't know if that's a bug or a feature, but I think we would have to deal with it somehow, right? What do you think @mariosky @pacastillo @ccottap @fergunet?