Leaving it for today
Up to the results section, pretty much finished. But please @mariosky @fergunet @unintendedbear have a look and revise fof badly written things, things missing, and things that shouldn't be there.

Main thing now is to insert new results, which include different cache sizes and the new problem, with 50 traps, in the loop. There is space for one page, but conclusions will be reduced, so I think there's enough. Will do it tomorrow, but please all pitch in, because tomorrow is the last day!