Rewrite abstract
I have done several changes, mainly on the client, as well as extensive experiments. Main change now is that overlap is almost entirely eliminated, that is, it is quite improbable that the clients started in a previous experiment contribute to a new one. However, this also makes changes in behavior of the users: they see changes if they leave it long enough. So let's see what goes on.

So far only thing changed is title and abstract. Please @mariosky @unintendedbear @fergunet take a look