Well, time for this one
I wasn't going to do it, but this is published as a TechRep and I think we should give it a try for EvoSoft.

Please @unintendedbear @fergunet @pacv @mariosky give it a shot more or less now. I'll submit a first version anyways and send another version when it is ready. I have to update it with the latest changes to the framework, for instance.