Addressing reviewer 3 comment
2) This part doesn't make sense:

"In general, a distributed volunteer-based evolutionary computation
system based on the browser is simply a client-server system which
client is embedded in the browser via JavaScript. We call this system
NodIO."

It's either a general statement, or it's a system called NodIO. It can't
be both. Please tidy up the wording.