R4 Typos
Typos:

Section 4
Paragraph 1. 'Since the point of using volunteer' -> 'Since the point
of using a volunteer'
Paragraph 2. 'to the rest, specially' -> 'to the rest, especially'
Paragraph 5. 'The first thin we will' -> ''The first thing we will'
Paragraph 6. ''which so needs less' -> 'which needs less'

Section 5
Paragraph 1. 'This configuration achieve a' -> 'This configuration
achieved a'